+++ src/app/app.module.ts
@@ -23,7 +23,7 @@ import { EllipsisPipe } from './pipes/ellipsis.pipe';
 import { KeysPipe } from './pipes/keys.pipe';
 import { SubsetPipe } from './pipes/subset.pipe';
 import { TokenPipe } from './pipes/token.pipe';
-import { RaidenConfig } from './services/raiden.config';
+import { RaidenConfig, Web3Factory } from './services/raiden.config';
 import { RaidenInterceptor } from './services/raiden.interceptor';
 import { RaidenService } from './services/raiden.service';
 import { SharedService } from './services/shared.service';
@@ -41,6 +41,7 @@ import { DragUploadDirective } from './directives/drag-upload.directive';
 import { StatusPipe } from './pipes/status.pipe';
 import 'hammerjs';
 import { FileUploadComponent } from './components/file-upload/file-upload.component';
+import { ErrorComponent } from './components/error/error.component';
 
 const appRoutes: Routes = [
     { path: '', redirectTo: '/home', pathMatch: 'full' },
@@ -85,7 +86,8 @@ export function ConfigLoader(raidenConfig: RaidenConfig) {
         AddressBookItemComponent,
         DragUploadDirective,
         StatusPipe,
-        FileUploadComponent
+        FileUploadComponent,
+        ErrorComponent
     ],
     imports: [
         RouterModule.forRoot(appRoutes),
@@ -119,7 +121,8 @@ export function ConfigLoader(raidenConfig: RaidenConfig) {
         },
         RaidenService,
         TokenPipe,
-        LocalStorageAdapter
+        LocalStorageAdapter,
+        Web3Factory
     ],
     entryComponents: [
         RegisterDialogComponent,