+++ src/app/services/shared.service.spec.ts
@@ -1,4 +1,4 @@
-import { TestBed, inject } from '@angular/core/testing';
+import { inject, TestBed } from '@angular/core/testing';
 
 import { SharedService } from './shared.service';
 
@@ -9,7 +9,14 @@ describe('SharedService', () => {
         });
     });
 
-    it('should ...', inject([SharedService], (service: SharedService) => {
+    it('should create', inject([SharedService], (service: SharedService) => {
         expect(service).toBeTruthy();
     }));
+
+    it('should have undefined status by default', inject(
+        [SharedService],
+        (service: SharedService) => {
+            expect(service.getStackTrace()).toBe(null);
+        }
+    ));
 });